+++ Changelog.md
@@ -1,3 +1,5 @@
 # Changelog
+* o [TASK] Load template based on selected backend layout
+---
 * ✓ [TASK] Load simple defaut template
 * ✓ [TASK] Create basic template structure